agent/requirements.txt
# 核心依赖
Flask==2.3.3
Werkzeug==2.3.7

# 加密和安全
cryptography==41.0.7
PyJWT==2.8.0

# 网络和通信
requests==2.31.0
urllib3==2.0.7

# 系统操作
psutil==5.9.6

# 配置管理
PyYAML==6.0.1

# 开发工具
black==23.9.1
flake8==6.1.0
mypy==1.6.1
isort==5.12.0

# 测试
pytest==7.4.3
pytest-cov==4.1.0

# 打包工具
pyinstaller==5.13.2